src/requirements.txt
Django==1.6.11
gunicorn==19.3.0
greenlet==0.4.5
gevent==1.0.2
celery==3.1.17
flower==0.8.2
requests==2.6.0
redis==2.9.0
beautifulsoup4==4.3.2
Fabric==1.10.1
python-simple-hipchat==0.3.3
librabbitmq==1.6.1
pika==0.10.0
